branch initialization

+++ src/main/java/com/nhnacademy/edu/springboot/account/controller/AccountController.java
@@ -1,17 +1,22 @@
 package com.nhnacademy.edu.springboot.account.controller;
 
-import com.nhnacademy.edu.springboot.account.domain.Account;
-import com.nhnacademy.edu.springboot.account.service.AccountService;
+
+import com.nhnacademy.edu.springboot.account.domain.AccountIdDTO;
+import com.nhnacademy.edu.springboot.account.entity.Account;
+import com.nhnacademy.edu.springboot.account.service.AccountClientService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
 public class AccountController {
-    private final AccountService accountService;
+    private final AccountClientService accountService;
 
-    public AccountController(AccountService accountService) {
+    @Autowired
+    public AccountController(AccountClientService accountService) {
         this.accountService = accountService;
     }
 
@@ -27,14 +32,29 @@ public class AccountController {
 
     @PostMapping("/accounts")
     @ResponseStatus(HttpStatus.CREATED)
-    public Account createAccount(@RequestBody Account account) {
-        return accountService.createAccount(account);
+    public ResponseEntity<AccountIdDTO> createAccount(@RequestBody Account account) {
+        try {
+            accountService.createAccount(account);
+        } catch (RuntimeException e) {
+            ResponseEntity.internalServerError().build();
+        }
+        return ResponseEntity.status(HttpStatus.CREATED).body(accountService.getAccountIdDTO(account));
     }
 
     @DeleteMapping("/accounts/{id}")
-    public String deleteAccount(@PathVariable Long id) {
-        accountService.deleteAccount(id);
-        return "{\"result\" : \"ok\"}";
+    public ResponseEntity<String> deleteAccount(@PathVariable Long id) {
+
+        if (accountService.exists(id)) {
+            try {
+                accountService.deleteAccount(id);
+            } catch (RuntimeException e) {
+                return ResponseEntity.internalServerError().build();
+            }
+            accountService.deleteAccount(id);
+            return ResponseEntity.ok().build();
+        } else
+            return ResponseEntity.notFound().build();
+
     }
 
 }
